I-#238: Fixed unittest issues
